fixed OpenGL using doubles

+++ app/shaders/vertex.glsl
@@ -1,12 +1,12 @@
 #version 460 core
 
-layout (location = 0) in vec2 v_position;
-layout (location = 1) in vec2 v_uv;
+layout (location = 0) in dvec2 v_position;
+layout (location = 1) in dvec2 v_uv;
 
-uniform mat3 u_transformation;
+uniform dmat3 u_transformation;
 
-uniform vec2 u_uvMultiplier;
-uniform vec2 u_uvOffset;
+uniform dvec2 u_uvMultiplier;
+uniform dvec2 u_uvOffset;
 
 out vec2 f_uv;
 
@@ -14,6 +14,6 @@ void main()
 {
 	gl_Position = vec4(u_transformation * vec3(v_position, 1.0f), 1.0f);
 
-	f_uv.x = u_uvOffset.x + (v_uv.x * u_uvMultiplier.x);
-	f_uv.y = u_uvOffset.y + (v_uv.y * u_uvMultiplier.y);
+	f_uv.x = float(u_uvOffset.x + (v_uv.x * u_uvMultiplier.x));
+	f_uv.y = float(u_uvOffset.y + (v_uv.y * u_uvMultiplier.y));
 }
